feat: Changes sets cards list response to prevent duplicate cards
Changes the data structure for the /sets/:code route so that is returns a list of all the card but regroup prints for a same card with a different rarity. The rarity attribute is now an array containing each possibility, instead of having multiple times the same card with only rarity to be different.